fix(plugin-openclaw): keep distinct fallback credentials, re-read a rotated config token
Review round 26, layer 4.

**Dedupe dropped a distinct credential (codex P1).** When a gateway-store token
wins for both, two configs on one endpoint resolve the SAME primary token, so
an endpoint+token key discarded the second candidate along with its different
bound credential - and the daemon's real token was never tried. The bound
credential is part of the identity now.

**Frozen config token (codex P2).** Copying the successful config credential
into `daemonAuthTokenOverride` meant every later request replayed a value that
a rotation invalidates, 401-ing every route until the gateway restarted. A
config-sourced credential is now SIGNALLED (`daemonAuthPrefersConfig`) and
re-read from `daemonConfigPath` per request; only a unit-supplied token, which
exists nowhere else, is still carried by value.

Two extractions keep both files under the 1200-line cap instead of raising the
ratchet: `delegate-authorization.ts` (the daemon authorization probe) and
`bridge-health-worker.ts` (the worker-thread probe body, which takes everything
by parameter and so cannot form an import cycle).
