+++ packages/plugin-openclaw/src/delegate-runtime.ts
@@ -25,6 +25,20 @@ import {
   renderMemoryContextPrompt,
 } from "@remnic/core";
 import { log } from "@remnic/core/logger";
+
+import {
+  DEFAULT_DELEGATE_AUTHORIZATION_OPERATIONS,
+  probeDelegateAuthorization,
+  reportDaemonAuthorizationFailure,
+  type DelegateAuthorizationOperation,
+  type DelegateAuthorizationPreflight,
+} from "./delegate-authorization.js";
+
+// Re-exported so existing importers of the runtime keep resolving.
+export {
+  probeDelegateAuthorization,
+  type DelegateAuthorizationPreflight,
+} from "./delegate-authorization.js";
 import {
   SESSION_NAMESPACE_BINDING_MAX_ENTRIES,
   SESSION_NAMESPACE_BINDING_MAX_NAMESPACES,
@@ -45,6 +59,7 @@ import {
   checkDaemonHealthSync,
   daemonUrl,
   loadDaemonAuth,
+  readDaemonConfigAuthToken,
   parseOpenClawBridgeConfig,
   resolveBridgeMode,
   resolveRequestedBridgeMode,
@@ -130,74 +145,6 @@ export interface DelegateHookApi extends DelegateCapabilityApi {
 }
 const DELEGATE_BATCH_FLUSH_CACHE_TTL_MS = 30_000;
 
-const DEFAULT_DELEGATE_AUTHORIZATION_OPERATIONS = [
-  "recall",
-  "observe",
-  "lcm_compaction_flush",
-  // The daemon-backed memory-slot capability searches through
-  // /engram/v1/memories/search, which enforces its own `memory_search`
-  // operation. Omitting it here would let the preflight report a
-  // least-privilege token authorized while every capability search 403s.
-  "memory_search",
-] as const;
-
-type DelegateAuthorizationOperation = (typeof DEFAULT_DELEGATE_AUTHORIZATION_OPERATIONS)[number];
-
-export interface DelegateAuthorizationPreflight {
-  readonly state: "authorized" | "unauthorized" | "unavailable";
-  readonly tokenSource: DaemonAuthToken["source"];
-  readonly status?: 401 | 403;
-}
-
-const daemonAuthFailureLogKeys = new Set<string>();
-
-function reportDaemonAuthorizationFailure(
-  serviceId: string,
-  pathname: string,
-  status: 401 | 403,
-  tokenSource: DaemonAuthToken["source"],
-): void {
-  const key = `${serviceId}:${pathname}:${status}:${tokenSource}`;
-  if (daemonAuthFailureLogKeys.has(key)) return;
-  daemonAuthFailureLogKeys.add(key);
-  log.error(
-    `delegate ${pathname} authorization failed (${status}; token source: ${tokenSource})`,
-  );
-}
-
-const AUTHORIZATION_PROBE_TIMEOUT_MS = 2_000;
-
-export async function probeDelegateAuthorization(
-  target: DelegateDaemonTarget,
-  namespace = "",
-  operations: readonly DelegateAuthorizationOperation[] = DEFAULT_DELEGATE_AUTHORIZATION_OPERATIONS,
-  /** Cap from a caller's shared deadline; the default is this probe's own. */
-  timeoutMs?: number,
-): Promise<DelegateAuthorizationPreflight> {
-  const auth = target.resolveAuthToken();
-  const headers = auth.token ? { Authorization: `Bearer ${auth.token}` } : undefined;
-  const query = new URLSearchParams();
-  for (const operation of operations) query.append("op", operation);
-  query.set("namespace", namespace);
-  try {
-    const response = await fetch(daemonUrl(target, `/engram/v1/authorization?${query}`), {
-      headers,
-      signal: AbortSignal.timeout(
-        timeoutMs === undefined ? AUTHORIZATION_PROBE_TIMEOUT_MS : Math.min(AUTHORIZATION_PROBE_TIMEOUT_MS, timeoutMs),
-      ),
-    });
-    await response.body?.cancel();
-    if (response.status === 200) {
-      return { state: "authorized", tokenSource: auth.source };
-    }
-    if (response.status === 401 || response.status === 403) {
-      return { state: "unauthorized", status: response.status, tokenSource: auth.source };
-    }
-  } catch {
-    return { state: "unavailable", tokenSource: auth.source };
-  }
-  return { state: "unavailable", tokenSource: auth.source };
-}
 async function postJson(
   target: DelegateDaemonTarget,
   serviceId: string,
@@ -1141,10 +1088,22 @@ export function maybeRegisterDelegateRuntime(
     // Bound to the config the resolved endpoint came from, so a deployment
     // with two configs sends each daemon its OWN token instead of whichever
     // one discovery happened to read first.
-    resolveAuthToken: () =>
-      bridge.daemonAuthTokenOverride === undefined
-        ? loadDaemonAuth(bridge.daemonConfigPath)
-        : { token: bridge.daemonAuthTokenOverride, source: "daemon configuration" },
+    resolveAuthToken: () => {
+      // A unit-supplied credential exists nowhere else, so it is used as-is.
+      if (bridge.daemonAuthTokenOverride !== undefined) {
+        return { token: bridge.daemonAuthTokenOverride, source: "daemon configuration" };
+      }
+      // A config-supplied one is RE-READ every request: freezing the value
+      // detection succeeded with would 401 every route after a rotation until
+      // the gateway restarted.
+      if (bridge.daemonAuthPrefersConfig && bridge.daemonConfigPath !== undefined) {
+        const configToken = readDaemonConfigAuthToken(bridge.daemonConfigPath);
+        if (configToken !== undefined) {
+          return { token: configToken, source: "daemon configuration" };
+        }
+      }
+      return loadDaemonAuth(bridge.daemonConfigPath);
+    },
   };
   registerDelegateRuntime(api, {
     serviceId: options.serviceId,
